created Appointment Action component, need to add status check and error messages
