+++ src/applications/check-in/pages/CheckIn/AppointmentListItem.jsx
@@ -1,41 +1,11 @@
-import React, { useState } from 'react';
+import React from 'react';
 import format from 'date-fns/format';
 import AppointmentLocation from '../../components/AppointmentLocation';
 
-import { api } from '../../api';
-
-import { goToNextPage, URLS } from '../../utils/navigation';
-
-import recordEvent from 'platform/monitoring/record-event';
+import AppointmentAction from '../../components/AppointmentAction';
 
 export default function AppointmentListItem(props) {
-  const [isCheckingIn, setIsCheckingIn] = useState(false);
-
   const { appointment, isLowAuthEnabled, token, router } = props;
-  const onClick = async () => {
-    recordEvent({
-      event: 'cta-button-click',
-      'button-click-label': 'check in now',
-    });
-    setIsCheckingIn(true);
-    try {
-      const checkIn = isLowAuthEnabled
-        ? api.v1.postCheckInData
-        : api.v0.checkInUser;
-
-      const json = await checkIn({
-        token,
-      });
-      const { status } = json;
-      if (status === 200) {
-        goToNextPage(router, URLS.COMPLETE);
-      } else {
-        goToNextPage(router, URLS.ERROR);
-      }
-    } catch (error) {
-      goToNextPage(router, URLS.ERROR);
-    }
-  };
 
   const appointmentDateTime = new Date(appointment.startTime);
   const appointmentTime = format(appointmentDateTime, 'h:mm aaaa');
@@ -58,16 +28,12 @@ export default function AppointmentListItem(props) {
           <AppointmentLocation appointment={appointment} />
         </dd>
       </dl>
-      <button
-        type="button"
-        className="usa-button usa-button-big"
-        onClick={onClick}
-        data-testid="check-in-button"
-        disabled={isCheckingIn}
-        aria-label="Check in now for your appointment"
-      >
-        {isCheckingIn ? <>Loading...</> : <>Check in now</>}
-      </button>
+      <AppointmentAction
+        appointment={appointment}
+        isLowAuthEnabled={isLowAuthEnabled}
+        router={router}
+        token={token}
+      />
     </li>
   );
 }
